Fix some UB cases. Closes #546

+++ soem/ethercatmain.c
@@ -349,7 +349,7 @@ uint8 ecx_siigetbyte(ecx_contextt *context, uint16 slave, uint16 address)
    {
       mapw = address >> 5;
       mapb = (uint16)(address - (mapw << 5));
-      if (context->esimap[mapw] & (uint32)(1 << mapb))
+      if (context->esimap[mapw] & (1U << mapb))
       {
          /* byte is already in buffer */
          retval = context->esibuf[address];
@@ -380,7 +380,7 @@ uint8 ecx_siigetbyte(ecx_contextt *context, uint16 slave, uint16 address)
          for(lp = 0 ; lp < cnt ; lp++)
          {
             /* set bitmap for each byte that is read */
-            context->esimap[mapw] |= (1 << mapb);
+            context->esimap[mapw] |= (1U << mapb);
             mapb++;
             if (mapb > 31)
             {
